Invoke Thor CLI in this ruby process. Instead of duplicating code and having output consistent, call the Thor check command for now. Keeping this simple
